// src/app/page/master/kecamatan/kecamatan.interface.ts
import { EventEmitter } from '@angular/core';

export interface KecamatanInterface {
  kode?: any;
  kdPropinsi?: string;
  kdkotaKabupaten?: string;
  kodeExternal?: string;
  namaExternal?: string;
  namaKecamatan?: string;
  reportDisplay?: string;
  statusEnabled?: boolean;
  noRec?: string;
  label?: any;
  value?: any;
  namaPropinsi?: string;
  id?: any;
  version?: any;
  propinsi?: string;
  kotaKabupaten?: string;
}